frontend: extract `getSelectedIndex` fn into a util file
